Create wizard utility; include some test scenarios;
Wizard utility interprets a “procedure”, which is just an array of objects of three members:  `conditions`, `step`, and `routing`. Conditions are another object whose keys may be names of the model properties against which conditional tests are being run against, and whose values are various conditional tests in the form of functions.